feat: implement GrpcStorageImpl#copy
Refactoring by pushing current CopyWriter implementation down to new
HttpCopyWriter and making CopyWriter abstract.
